lsass:tools: Fix output of 64-bit timestamp in get account/password tool.
CR: aberry@likewise.com

git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@53882 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lsass/tools/lw-lsa/ad_get_machine.c
@@ -32,7 +32,7 @@
  *
  * Module Name:
  *
- *        FILENAME
+ *        ad_get_machine.c
  *
  * Abstract:
  *
@@ -151,7 +151,7 @@ PrintAccountInfo(
            "  FQDN: %s\n"
            "  Join Type: %u\n"
            "  Key Version: %u\n"
-           "  Last Change Time: %d\n"
+           "  Last Change Time: %ld\n"
            "",
            LW_PRINTF_STRING(pAccountInfo->DnsDomainName),
            LW_PRINTF_STRING(pAccountInfo->NetbiosDomainName),
@@ -160,7 +160,7 @@ PrintAccountInfo(
            LW_PRINTF_STRING(pAccountInfo->Fqdn),
            pAccountInfo->Type,
            pAccountInfo->KeyVersionNumber,
-           (int) pAccountInfo->LastChangeTime);
+           pAccountInfo->LastChangeTime);
 }
 
 static
